# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add_or_delete_in_favorite": {
            "post": {
                "tags": [
                    "Favorite"
                ],
                "summary": "add_or_delete_in_favorite",
                "description": "add_or_delete_in_favorite",
                "operationId": "add_or_delete_in_favorite",
                "requestBody": {
                    "description": "Pass user register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "product_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "deleted is Favorite",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Created Is Favorite",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/get_my_favorite": {
            "get": {
                "tags": [
                    "Favorite"
                ],
                "summary": "get_my_favorite",
                "description": "get_my_favorite",
                "operationId": "get_my_favorite",
                "requestBody": {
                    "description": "Pass user register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "deleted is Favorite",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Created Is Favorite",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}